requirements.txt
numpy>=1.26
scipy>=1.11

pandas>=2.1

matplotlib>=3.8
seaborn>=0.13

pyyaml>=6.0

gymnasium>=0.29
stable-baselines3>=2.2

tqdm>=4.66

torch